not calc rtf and ltc

+++ egs2/vlsp/asr1_streaming/asr.sh
@@ -60,7 +60,7 @@ min_wav_duration=0.1 # Minimum duration in second.
 max_wav_duration=20  # Maximum duration in second.
 
 # Tokenization related
-token_type=bpe      # Tokenization type (char or bpe).
+token_type=char      # Tokenization type (char or bpe).
 nbpe=30             # The number of BPE vocabulary.
 bpemode=unigram     # Mode of BPE (unigram or bpe).
 oov="<unk>"         # Out of vocabulary symbol.
@@ -126,7 +126,7 @@ nll_batch_size=100 # Affect GPU memory usage when computing nll
                    # during nbest rescoring
 k2_config=./conf/decode_asr_transformer_with_k2.yaml
 
-use_streaming=false # Whether to use streaming decoding
+use_streaming=true # Whether to use streaming decoding
 
 use_maskctc=false # Whether to use maskctc decoding
 
@@ -1586,7 +1586,7 @@ if [ ${stage} -le 12 ] && [ ${stop_stage} -ge 12 ] && ! [[ " ${skip_stages} " =~
                 ${_opts} ${inference_args} || { cat $(grep -l -i error "${_logdir}"/asr_inference.*.log) ; exit 1; }
 
         # 3. Calculate and report RTF based on decoding logs
-        if [ ${asr_task} == "asr" ]; then # sửa chổ này cho nó tính RTF và latency xem nha
+        if [ ${asr_task} == "asr" ] && [ -z ${inference_bin_tag} ]; then # đã tắt tính RTF sửa chổ này cho nó tính RTF và latency xem nha
             log "Calculating RTF & latency... log: '${_logdir}/calculate_rtf.log'"
             rm -f "${_logdir}"/calculate_rtf.log
             _fs=$(python3 -c "import humanfriendly as h;print(h.parse_size('${fs}'))")
